Move sub classes into their own files and change presenter and view types for Supplement.cs
